Handle the case where RELEASE SAVEPOINT fails with a retry
From the CockroachDB Docs:

In some cases, the RELEASE SAVEPOINT statement itself can fail with a
retry error, mainly because transactions in CockroachDB only realize
that they need to be restarted when they attempt to commit. If this
happens, the retry error is handled as described in step 4.

This commit includes the RELEASE SAVEPOINT in the retry-able part of the
transaction.