introduce ParseRelationships
given a relationships block from the validation file,
returns a deduplicated slice of RelationTuple.

Given this logic was already present in loader.go, I attempted
to DRY it. In order to achieve that and retain the logic
in loader.go, the new function needs to be able to parse
both newline-separated blocks (present in the relationship block
in the YAML) or individual tuple strings, as present in the "validation" block