introduce validate command
the purpose of this command is to take a playground file and run
the assertions and validations defined.

The rationale is that schema development happens in the playground,
but once the YAML is downloaded, there is nothing developers can do
other than loading it with testserve command, or uploading it back to the playground.
This attempts to reuse and run the assertions and validations as test-suite outside of the
the playground, and in a programmatic way rather than only interactively.
Rather than duplicating the same tests in the client application, the playground tests
become the canonical representation for the business rules defined in the schema.

Example:
- developers introduce changes in schema via the playground
- YAML file is downloaded and persisted in git repository
- changes are pushed, PR is opened, CI runs spicedb validate, demonstrating changes are sound.