add a cap to iterations for newenemy test
occasionally the number of iterations picked will take longer than the
timeout for the test suite. capping it means that we sometimes will not
run the full iterations to be convincing, but the effect should be
amortized over PR runs. The cap (and timeout) can be removed for any
change that we suspect may have implications for the test.