SECURITY FIX: Ensure wildcard is properly handled in Lookup
The Lookup design called for wildcard to be ignored, but this is not doable since wildcard can be found on the *right* side of an exclusion or within an intersection, which means lookup must handle those case. Without it, a Lookup could return an item that would otherwise Check as false, which is a permissions model inconsistency.

This change ensures lookup handles following paths from the wildcard of the subject as well (where relevant), enables consistency testing of those cases, and adds additional testing to ensure all assertions are performed against lookup (in addition to check).