dispatcher: fetch datastore from context
the dispatcher is almost always used in places where we have a
middleware injecting the datastore (and it's easy to inject in places
we don't, like in tests).

fetching the datastore from the context means that a single dispatcher
instance can be shared between multiple datastores; i.e. for the
testserver