move DB specifics into its package
each one of mysql and postgresql transaction, transaction
beginner and rows appeared in a "common" package, which
leaks implementation details.

After moving them to the corresponding package, I realized
that transaction.go was under a "rdb" package, but
the code in "common" used it, so it didn't make sense
to have it in its own package since they are already coupled.