fixes mysql datastore flakes
since the introduction of the instrumented mysql connector,
some errors tests have been emerging caused by the registration
of a duplicated metric.

In order to address it, I made the instrumentation only enabled
when the corresponding prometheus option is enabled, which also
does makes sense to have all metrics controlled by that flag.
Since we only have 1 test that enables prometheus, metrics
this guarantees we don't register the metric a second time.