introduce way to run with backward compatible revisions
the IsReady method always checks that the revision
in the datastore is exactly the same as HEAD. This is problematic
when running migrations because it's not possible to steplock
the execution of a migration and deployment of the new codebase that
requires it.

- if you run the migration, IsReady() will start returning false and cause downtime,
because the code running detects a revision that isn't expected
- if you deploy the code first, IsReady will also start returning false and cause downtime,
because the expected migration is not run yet.

the proposed solution is to allow the datastore running with revisions that are
"backward compatible with head". the assumption is the revision replaced by head is always compatible.
This makes it possible to execute the following workflow:

- deploy new version of code that, making sure it's capable of running with previous version as well
- run migration
- clean up backward compatibility code on the next migration introduced

The alternative would be to allow the codebase running with an "unknown future" migration, and make
the assumption that an unknown revisions can be ignored.