reduces SQL queries in the critical path for optimized revisions
this change attempts to make optimizedRevision faster by
avoiding one SQL query, going from 3 down to 2.

While these are fast queries, the bottleneck
is the network and the connection pool size. By removing one query
we spend less time waiting for the network and for connections
becoming available in the pool.