Implement a reachability graph
The reachability graph is a helper system for computing the entrypoint relations/permissions for a subject of a specific type when trying to *reach* a target relation/permission. This will form the basis of the reachable resources dispatch and API, and hopefully be used to significantly narrow down the working set for lookup in the future.