introduces mysql migration driver
GitHub runs mysql schema migrations at scale using skeema
(https://github.com/skeema/skeema) and gh-ost (https://github.com/github/gh-ost).
The SpiceDB migration driver for MySQL is designed in a way it
is compatible with migration workloads relying on those tools.

The main difference is how the MySQL migration driver maintains
the metadata needed to know which version is currently running. The
PSQL migration driver follows a pattern similar to what's proposed
by SQLAlchemy's Alembic, with a table that keeps track of the migration run.
Because Skeema does not support data migrations, and seeding the database
could be considered a data migration, we decided to encode the metadata
into the database schema (DML) so we can expose the same information
but only relying on schema migrations. As such the migration_version table
is empty, and maintains the running version in a table column. This turns
every migration into a pure DML operations, which Skeema understands well.

The way we run these then is by executing zed migrate in development DB,
and letting skeema detect the changes to the development database. These are
then persisted, and deployed using our own internal migration automation
buil on top of skeema and gh-ost.

The other distiction with the PSQL migration driver is the support for table
prefixes. This is necessary to align with our own internal conventions.

Co-authored-by: Alex Slepak <ams11@github.com>
Co-authored-by: Christopher Kirkland <chriskirkland@github.com>
Co-authored-by: Hannah Gould <hagould@github.com>
Co-authored-by: John Reed <johnpreed@github.com>
Co-authored-by: Sean Bryant <sbryant@github.com>