pr feedback: clarify comment
This should theoretically not be necessary, but while learning on how MySQL did this,
we found some behaviour we couldn't quite decipher.

- MySQL is governed by a global timezone parameter.
- DATETIME is stored with timezone information.
- sessions can also have their own timezone and override the server global
- when we query the RelationTupleTransaction, the driver is able to obtain
  DATETIME honoring the timezone it was stored with. The parseTime
  flag in the go driver enables properly unmarshalling the query response into a time.Time
- However, when performing a query without a table like SELECT NOW(), somewhere in the way
  the timezone wasn't being recognized. We spent some time debugging the driver but
  couldn't find the answer. With some tweaks to the TestTransactionTimestamps we managed to
  reproduce the problem, and fixed it by forcing the timezone in the query SELECT UTC_TIMESTAMP().

We ended up leaving it this way because we found it easier to convert all time.Time instances flowing
through SpiceDB to UTC just to make it easier to visually inspect it while debugging, so the code was
here for a different reason that the PSQL driver.