introduces mysql datastore (#525)
* introduces mysql migration driver

GitHub runs mysql schema migrations at scale using skeema
(https://github.com/skeema/skeema) and gh-ost (https://github.com/github/gh-ost).
The SpiceDB migration driver for MySQL is designed in a way it
is compatible with migration workloads relying on those tools.

The main difference is how the MySQL migration driver maintains
the metadata needed to know which version is currently running. The
PSQL migration driver follows a pattern similar to what's proposed
by SQLAlchemy's Alembic, with a table that keeps track of the migration run.
Because Skeema does not support data migrations, and seeding the database
could be considered a data migration, we decided to encode the metadata
into the database schema (DML) so we can expose the same information
but only relying on schema migrations. As such the migration_version table
is empty, and maintains the running version in a table column. This turns
every migration into a pure DML operations, which Skeema understands well.

The way we run these then is by executing zed migrate in development DB,
and letting skeema detect the changes to the development database. These are
then persisted, and deployed using our own internal migration automation
buil on top of skeema and gh-ost.

The other distiction with the PSQL migration driver is the support for table
prefixes. This is necessary to align with our own internal conventions.

Co-authored-by: Alex Slepak <ams11@github.com>
Co-authored-by: Christopher Kirkland <chriskirkland@github.com>
Co-authored-by: Hannah Gould <hagould@github.com>
Co-authored-by: John Reed <johnpreed@github.com>
Co-authored-by: Sean Bryant <sbryant@github.com>

* test driver registering invalid version names

* roll back to 1.5.0 MySQL driver

some recent changes to telemetry seems to be failing
with the 1.6.0 MySQL Driver version. 1.5.0 was an
indirect dependency already in the main branch.

* introduces mysql datastore implementation

in its current state the MySQL datastore duplicates a
significant portion of the PSQL datastore's logic. Given
the fast pace of breaking changes in the internal implementation,
we considered code should be upstreamed ASAP and start addressing
duplication and gaps as more-approachable and smaller follow ups.

The fundamental differences with the PSQL implementation are:
- usage of Go's databases/sql API
- queries need to be generated on Datastore initialization
  in order to handle customizable prefix functionality
- transaction seed in runtime: the implementation tries
  hard to avoid data migrations. Instead of handling the
  seeding of "transaction zero" as part of "spicedb migrate",
  we introduced it in IsReady(). This will be revisited in a follow-up,
  to explore moving it to the datastore struct initialization.
  - SQL query changes: some queries were tweaked to make it compatible
  with Vitess
- IsReady(): introduces changes to avoid downtime during migrations.
  The implementation relies on a new function "IsHeadCompatible()", which
  has a rather simplistic heuristic: compatible if is HEAD or the version
  it HEAD replaced. This is probably the most controversial part of the change
  given it introduces a new method as part of the API.
- Stats cluster ID is not implemented - we didn't see an obvious need to introduce
  the complexity needed to handle it, but if controversial can be handled in a follow up
- only supports a subset of the CLI options. Those will be handled as part of a follow up
- does not emit metrics for garbage-collection - also to be handled in a follow up
- exposes connection pool metrics via dlmiddlecote/sqlstats

This commit does not wire the datastore into the spicedb CLI.

Co-authored-by: Alex Slepak <ams11@github.com>
Co-authored-by: Christopher Kirkland <chriskirkland@github.com>
Co-authored-by: Hannah Gould <hagould@github.com>
Co-authored-by: John Reed <johnpreed@github.com>
Co-authored-by: Sean Bryant <sbryant@github.com>

* fixes missing check rows.Err() after rows.Next()

this was caught in the mysql implementation by the linter.
For some reason the linter didn't detect it in the PSQL
implementation.

* fixes typo in span name

* pr feedback: put type close to methods

* pr feedback: reword struct comment

* pr feedback: cluster statements

* pr feedback: rename tag and invert logic

* pr feedback: style - new lines after conditionals

* pr feedback: style - new lines after conditionals

* pr feedback: move comment and elaborate on the design rationale

* pr feedback: naming consistency

* pr feedback: add comments to exported types

* pr feedback: change to TODO and add individual

* pr feedback: clarify comment

This should theoretically not be necessary, but while learning on how MySQL did this,
we found some behaviour we couldn't quite decipher.

- MySQL is governed by a global timezone parameter.
- DATETIME is stored with timezone information.
- sessions can also have their own timezone and override the server global
- when we query the RelationTupleTransaction, the driver is able to obtain
  DATETIME honoring the timezone it was stored with. The parseTime
  flag in the go driver enables properly unmarshalling the query response into a time.Time
- However, when performing a query without a table like SELECT NOW(), somewhere in the way
  the timezone wasn't being recognized. We spent some time debugging the driver but
  couldn't find the answer. With some tweaks to the TestTransactionTimestamps we managed to
  reproduce the problem, and fixed it by forcing the timezone in the query SELECT UTC_TIMESTAMP().

We ended up leaving it this way because we found it easier to convert all time.Time instances flowing
through SpiceDB to UTC just to make it easier to visually inspect it while debugging, so the code was
here for a different reason that the PSQL driver.

* pr feedback: newline

* pr feedback: refactor seed base transaction

the initial ask was to move the log message. While inspecting
further the method it became obvious there was no reason
to scatter the seeding logic and that everything done in IsReady
should be encapsulated in seedBaseTransaction().

* pr feedback: unintended extra //

* pr feedback: constantize magic number

* pr feedback: document and simplify with Sprintf

* pr feedback: style on DML statements

* fix broken links

* pr review: do proper Sprintf and separate arguments with commas and newlines

* pr feedback: consolidate into a single migration

* be explicit about engine and charset

if the database defaults to latin1, users
running the datastore may end up in trouble
when needing to store unicode characters. Having
it upfront will likely spare some pain in the future.

* pr feedback: document exported types

Co-authored-by: Alex Slepak <ams11@github.com>
Co-authored-by: Christopher Kirkland <chriskirkland@github.com>
Co-authored-by: Hannah Gould <hagould@github.com>
Co-authored-by: John Reed <johnpreed@github.com>
Co-authored-by: Sean Bryant <sbryant@github.com>