moves seeding to the initialization of the datastore
this moves the seeding responsibility out of IsReady
and makes it side-effect free. Having the mutation
as part of the datastore initialization gives us
some nice properties:
- the application fails to start if seeding didn't succeed. Lack
of seeding is not something the application can recover from,
and we wouldn't want to have a SpiceDB pod running as healthy
when the datastore isn't seeded. If seeding failed due to a transient,
the pod will get another chance as it's restarted by kubernetes.
- IsReady is not invoked right after SpiceDB starts, so SpiceDB
can start even though the datastore is not ready. This was
reported in https://github.com/authzed/spicedb/issues/482 without
traction on what was proposed, so by moving seeding to datastore initialization
we can make sure the application will fail to start.