introduces Transact function in migrate.Driver
the goal of this new function of the interface is to
be able to execute a migration operation transactionally.

Given each driver has different approaches to create a transaction,
generics are used to expose the transactional boundary to each
migration function without going attempting to create transaction
interface that abstracts all datastores.