move transacting responsibility to the underlying driver
this design is the response to the fact that datastores
cannot solely run MigrationFuncs based on the corresponding
transaction type.
- Spanner: DML and DDL operations cannot run transactionally
- CRDB: "SET CLUSTER SETTINGS" cannot be run inside a transaction
- PostgreSQL: CREATE INDEX CONCURRENTLY cannot run inside a transaction block (SQLSTATE 25001)

The initial attempt to fix this introduced 2 different method in the Driver interface, one for
transactional migration functions, and another for non-transactional migration functions. Because
of the use of generics this effectively led to duplicating code to handle the 2 types.

Ultimately I realized that the source of the complexity was coming from forcing the Manager
to invoke the "Transact", that imposed transactional semantics in the underlying driver. Instead
this new design completely hands over to the MigrationFunc, who shall decide the type of semantics
needed. This means MigrationFunc is now responsible to:
- defining transactional semantics when needed
- making sure the version is written

This seemed like a reasonable tradeoff and does not impose too much of a burden to the driver implementation.
As it can be seen from the diff, most of the overhead comes from making sure each MigrationFunc writes
the corresponding version. Not only this can be abstracted away with a helper method specific to each
datastore, but the Manager can also do a runtime check after the function is executed to make sure the
datastore has effectively transited to the new version, as a mechanism to avoid human mistakes.