removes v1.RelationshipUpdate from datastore.ReadWriteTransaction
this refactor replaces v1.RelationshipUpdate with core.RelationTupleUpdate
in the datastore.ReadWriteTransaction interface. The goal is to
help decouple the datastore from the public interfaces and use internal
data structures to help support implementation details, like those
needed for Caveats. This makes sure we don't need to push implementation
details to the public interface just because it's needed in the datastore

this commit is rather large as I took the opportunity to reduce duplication.
I found mysel doing repetitive work during the refactor that could have been
lessened with some DRY.

I added a temporal workaround to TestDeleteRelationship because I found datastores aren't able
to cope with duplicate tuples. This didn't surface earlier because most tests
called ReadWriteTransaction.WriteRelationships multiple times with single-element
slices, instead of having all tuples in the slice. I didn't want this changeset to
blow up in scope, so I made sure the original unintentional behaviour remains.

Finally, ValidateUpdatesToWrite was removed from datastore/common package because it was
only used in testfixtures, and it called HandwrittenValidate. This is unavailable in core,
so I converted the core tuple to a v1 relationship just to call HandwrittenValidate. This
should be fine because this is just used for tests and not in the critical path.