Improve performance of direct checks on relations by using targeted SQL queries in supported datastores
After this change, when performing a direct check lookup on a relation, a constructed query will be used to try to minimize the number of relationships returned. For example, if a target subject is allowable on a relation and is a terminal subject, then *only* the actual subject itself will be queried, rather than looking for all subjects of that type and comparing fully in memory. This results in a significant performance boost when there are very wide relations, mostly filled with terminals, which is a common pattern of data.