adjust naming and use structpb
instead of using protobuf.Any, we use
protobuf.Struct that better maps a
map[string]any

we move away from the concept of caveat
digest and will reference them by name.
Anonymous caveats will have a globally-unique
name that may likely be the result of performing
a digest of the signature and payload

we replace "caveat logic" with "caveat expression"

we also change the term "predefined variables" with "caveat
context" to refer to the state persisted alongside the tuple.
these are variables that will be injected at caveat evaluation time

as a minor inconvenience the relationship.RelationTuple method now has to
return an error, propagated from the call to structpb.NewStruct.
That led to changes in all callsites using it.