introduce caveat support in WriteRelationships
this commit changes parts of the datastore implementation.
The duality caveatID/CaveatName was problematic to handle
between the controller layer and the datastore layer:

The usage of ID introduced a situation where a core.RelationTuple
cannot be converted to a v1.Relationship without a database lookup,
because database tuples reference caveats via ID, but we return
named caveats to the client via v1.Relationship. It would be a departure
from what happens today in service controllers: v1-to-core would introduce
I/O operations.

The question here is that it's not clear the usage of ID buys us anything
at this point. The rationale behind them was being able to support "anonymous caveats".
An anonymous caveat signature defined in place in a relation in the schema.
Because they don't have a name, we need a way to identify them.

However, it's not clear anonymous caveats is needed at this point,
and that we should work on it when the need arises.