specifies cache-dependency-path to prevent poisoning
these modules have very little dependencies,
and end up "winning" in the race to cache.

By introducing the path we can have the Github Action
disambiguate the cache key name by path.

From the docs:
The action defaults to search for the dependency file -
go.sum in the repository root, and uses its hash as a
part of the cache key. Use cache-dependency-path input for
cases when multiple dependency files are used, or they are
located in different subdirectories.