Add trace debugging setting for use in the development API
When enabled, all batch operations in Check will be disabled, to ensure that the produced debug trace is as accurate as possible. As this option is only used against a memdb-based database where chunking provides little value, the performance overhead for the development API should be essentially nil.

This change also adds the produced debug information onto the development package, for use by other tooling