redesign query interception
because the PGXExecutor is not the only
entry point to perform database queries
that impact the critical path of SpiceDB,
this commit proposes wrapping the pgx.Pool
to add an interceptor layer. This way every
new use of the pooler will be intercepted and
we won't miss any potential future callsites.

It also allows us to build tests for methods
like HeadRevision / OptimizationRevision