redesigns middlware options for RunnableServer
the methods to change middleware in RunnableServer
were introduced because we needed a mechanism
to be able to retain the default middleware and
expand it.

This new design is explicit about it, by providing
different operations to change the middleware layer:
prepend, replace and append.

This allows us to remove the methods from RunnableServer
and configure the final middleware chain through the
server.Config