fix leaking resources when Complete() errors
if any part of Complete() fails, the method
returns without calling Close() on all the
resources that could have possibly
be successfully initialized. This is
problematic in testing, when we want
to start up servers and use cancellation
signals to gracefully terminate.

the approach here taken is to
collect all closers during the
execution of Complete(), and
add a defer statement that
calls them if an error happened.
The completedServerConfig.closeFunc
also changes to call each one of the closers.