redesign middleware modification API
this new design try to explore a more
explicit approach to modifying
the middleware chain. Any modification
will have to indicate where in the
chain they want the middleware to be
added.

The default SpiceDB middleware layer has
some middlewares that are not critical
for its functionality, but others that
need to be there. For that reason
middlewares can be marked as internal.
An internal middleware can be referenced on
append or prepend operations, but not on replace.