pkg/cache: implement a central collector
This commit adds a Prometheus collector that
records metrics for all allocated caches.

It fixes a bug where only the dispatch cache was
being reported because it was manually done in
the caching dispatcher.

This also changes the default to collect metrics.
which is the common case as the overhead is worth
the added observability.