fixes GC behaviour with out-of-window head revision
recent changes in https://github.com/authzed/spicedb/pull/1153
could lead to a situation where the transactions table is empty.

Without an anchor, it's not possible to determine if a provided
revision is valid, unless we change the revision implementation
in PG that provided the snapshot and the time.

This change undoes a change in https://github.com/authzed/spicedb/pull/1153
that will cause the last revision to be removed, and adjusts
queryValidTransaction so it return the last transaction recorded
as valid, even if it's outside of the GC window, because it's the
last known state of the system.

There is a scenario we couldn't reconcile in this commit:
currently the various datastores behave differently when a
when a request comes with a revision that was obtained one GC window ago.
this is due to the fact MySQL and PostgreSQL implement revisions as a snapshot,
while CRDB, Spanner and MemDB implement it as a timestamp.