fixes comeback of the MacOS microsecond precision problem
we recently changed how MemDB handles head revisions
that are outside of the window. We iterated
on the fix, but the final solution was the result
of merging 2 solutions, by accident.

- one option was to change HeadRevision to return now
- another option was to handle in CheckRevision and
  make an exception if the revision is the last known
  revision, even if outside of GC window

Both fixes work fine together, but the first option
resurfaced problems with time resolution in MacOS
that leads to test flakes. This happens because
a call to very close calls to HeadRevision and
ReadWriteTx could lead to the same revision, due to
the microsecond resolution.

The proposed solution is to rollback the changes
to HeadRevision. While we want to converge to time-based
revisions, we already have other datastores doing this
anyway, so addressing this would be part of a bigger
initiative.