Fix deduplication bug in reachable resources
We were previously deduplicating dispatching in reachable resources based solely on the subject found, but without being transformed for the parent resource type. This resulted in branches that produced the *same* subject, but with a different relation, being skipped occasionally due to a race in the dispatching.

We now do the duplication check at the dispatch point, to ensure we're checking the exact subject being dispatched.

Also adds a consistency test which reproduces the issue. Thanks to `tr33` from Discord for the bug report and reproduction case!