balancer: rewrite the consistent hashring balancer to avoid recomputates
The previous implementation was recomputing the hashring any time a
subconnection moved from ready->idle or back, which happened frequently.

The new implementation includes idle and connecting subconns in the ring,
and triggers a connection if one is selected. It also adds/removes from
a long-lived ring instead of recomputing a ring from scratch each time.

ReplicationFactor and Spread can now be passed in as part of the
service config instead of registered globally with the balancer