move to paths-filter build-skipping strategy
the problem with the usage of "paths" is that
is that it gets in conflict with required builds.
If a build is required, and skipped with "paths",
and passing checks are required (which is the case
here) then the PR will never go green and ready to
merge. This affects the same way the merge-queue.

This commit removes any usage of paths in favour
of the paths-filter GHA. What it does is to generate
determine if there was a change in certain paths,
and if that's the case, you can use "if:" option
to skip a job. Required builds can be skipped.