introduce termination.Error to support termination-log
termination log is a concept in Kubernetes that
allows applications providing information
upon termination that will be surfaced as part
of the status of a Pod. This is useful for troubleshooting
purposes but also can be used by the control plane.

The termination.Error also includes an ExitCode which
will when terminating the process.