more robust server defaults using optgen defaults
it was still possible to end up creating a server.Config
programatically that didn't have defaults set. This
causes constant friction when a new flag is added.

This introduces the usage of optgen defaults in
server.Config, and replaces any Config value creation
with the helper method that invokes `defaults.MustSet`