decouples middleware modifications
until now middleware modifications composed both unary and
streaming middlewares. This worked under the assumption
that we wanted the same order in each type of API method.

SpiceDB relies on `grpc-ecosystem/go-grpc-middleware` to build
[chains of interceptors](https://github.com/grpc-ecosystem/go-grpc-middleware/blob/bfde8086e68d6a129363b063ffd6518ab568467d/doc.go),
which is not natively
supported by the go gRPC SDK.
The order of execution is, assuming a slice of interceptors, from left to right.

Because of the design of gRPC Streaming APIs, composed of stream creation and actual
streaming of messages, while the chain of interceptors will follow the expected
middleware chain order, the actual interception of messages (`RecvMsg` and `SendMsg`)
will happen in the opposite order.

Having middlewares depend on each other is a not good practice, but there
are circumstances where at the very least you need a middleware to be the first
or the last in the chain, think middlewares for auditing purposes should happen
earlier, otherwise, it wouldn't catch requests that fail earlier in the chain.

This can be addressed by reversing the order of the middleware to obtain the
expected order, but this hasn't been possible because the `MiddlewareModification`
API had both unary and streaming coupled.

This commit separates modifications to unary and streaming middlewares, generalized using Go generics.

The generated options also rely on a modification to `ecordell/optgen` in order
to support generics. See https://github.com/ecordell/optgen/pull/9