Redo how context is handled around limits
The previous approach of having the limit tracker control the context was causing parallelized branches of the Reachable Resources to be canceled early; we were ignoring this error (incorrectly) and thus, reachable resources could return partial results.

Instead, we now manually and explicitly detach the context and cancel it explicitly when no further dispatching is needed.