fixes incorrect usage of backoff in GC
once `MaxElapsedTime` has elapsed, the backoff switches
to the value of `Stop` which is -1. This mean it
will effectively remove the backoff and a storm of requests
will hit the DB.

I also changed it in every instance I found because the default
behaviour is dangerous.

See https://github.com/authzed/spicedb/issues/1444