Change how dispatches are counted in LookupResources
Before the cursoring change, LR and RR issued *batch* operations, with the batch dispatch counting as only a single dispatch. Following the cursoring change, which changed all results to be streamed back individually, the dispatch was counted in *every* result, which caused the aggregation to be MUCH larger than correct in the LookupResources API handler.

This change adds tests to ensure that the dispatch count is in line with expectations when using batching, and fixes the above issue by only counting the dispatch once on the first result streamed back. If the first result (or any subsequently) are filtered, then the dispatch count is "held" until another result is ready to be published and the count can be appended to that result. This does mean that we might get *lower* counts, but never higher.